Convert `serve` command and action to mjs
