--- packages/sku/lib/program/options/port/port.option.js
@@ -1,4 +1,0 @@
-const { Option } = require('commander');
-const portOption = new Option('--port [port]', 'Port to serve on');
-
-module.exports = portOption;
