+++ packages/sku/lib/program/options/port/port.option.mjs
@@ -1,0 +1,3 @@
+import { Option } from 'commander';
+
+export const portOption = new Option('--port [port]', 'Port to serve on');
