--- packages/sku/lib/program/options/site/site.option.js
@@ -1,4 +1,0 @@
-const { Option } = require('commander');
-const siteOption = new Option('--site [site]', 'Site to serve on');
-
-module.exports = siteOption;
