+++ packages/sku/lib/program/options/site/site.option.mjs
@@ -1,0 +1,3 @@
+import { Option } from 'commander';
+
+export const siteOption = new Option('--site [site]', 'Site to serve on');
